Delete macosx.h
It was essentially identical to unix.h, and OS X is very Unix-like
(most of the newer versions are officially Unix-certified). However,
the __unix__ macro is not defined on OS X, so I added a check for
the __APPLE__ macro and define __unix__ based on that. The Unix code
should work on OS X without any problems.